changed Float64 to AbstractFloat, Int64 to Int and initPhi to initJacobi

+++ findall.jl
@@ -1,9 +1,9 @@
 import Base.findall
 function findall(x::Array{T,1} where T<:Real, v::Array{T,2} where T<:Real)
-    sol=Array{Int64,1}(undef,0)
+    sol=Array{Int,1}(undef,0)
     #length(x) != size(v,1) && error("Dimensionen passen nicht zusammen.")
 
-    (rows::Int64, cols::Int64)=size(v);
+    (rows::Int, cols::Int)=size(v);
     for i in 1:cols
         for j in 1:rows
             if !isapprox(x[j],v[j,i])
@@ -19,7 +19,7 @@ end
 
 
 function findall(x::Real, v::Array{T,1} where T<:Real)
-    sol=Array{Int64,1}(undef,0)
+    sol=Array{Int,1}(undef,0)
     for i in eachindex(v)
         if isapprox(v[i],x)
             push!(sol,i)
